social-analyzer/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fake-useragent==1.5.1
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
load-dotenv==0.1.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.25.0
packaging==24.0
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
sniffio==1.3.1
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.40.1
twikit==1.5.11
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
